% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_hypothesis_params}
\alias{calculate_hypothesis_params}
\title{Calculate Hypothesis Test Parameters}
\usage{
calculate_hypothesis_params(df, alpha = 0.05)
}
\arguments{
\item{df}{Data frame where each column is a vector of p-values from a statistical test}

\item{alpha}{Numeric significance level; defaults to 0.05}
}
\value{
A data frame with columns for Type 1 error, Type 2 error, and power as well as rows for each test
}
\description{
This function calculates the power, Type II error, and Type I error of tests given p-values
}
