% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{libname}

\item{pkgname}{package}
}
\value{
invisible
}
\description{
This function is called when the package is loaded. It sets the default options for the package.
}
\keyword{internal}
