% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ordinalsimr_options.R
\name{.set_options_helper}
\alias{.set_options_helper}
\title{Set Option Helper}
\usage{
.set_options_helper(option_name, value, additional_msg = NULL)
}
\arguments{
\item{option_name}{name of the ordinalsimr option to set}

\item{value}{value to set the option to}

\item{additional_msg}{additional message to print after setting the option}
}
\value{
invisible
}
\description{
Set Option Helper
}
\keyword{internal}
