% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_simulation_data}
\alias{format_simulation_data}
\title{Format Simulation Data}
\usage{
format_simulation_data(input)
}
\arguments{
\item{input}{a named list structured as the output from `run_simulations()`}
}
\value{
a named list with a p_value tibble and a run_info nested tibble
}
\description{
Soft wrapper for `dplyr::bind_rows()`: formats the list of data from running simulations into a list of tibbles; one tibble for the p-values for each run of each model, and one tibble with run metadata. List objects are named `p_values` and `run_info`. The tibble in `run_info` further contains two nested tibbles.
}
