% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ordinalsimr_options.R
\name{get_ordinalsimr_options}
\alias{get_ordinalsimr_options}
\title{Get ordinalsimr options}
\usage{
get_ordinalsimr_options()
}
\value{
list of ordinalsimr options
}
\description{
Returns all of the ordinalsimr options to the console.
}
\examples{
get_ordinalsimr_options()
}
