% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tests.R
\name{ordinal_tests}
\alias{ordinal_tests}
\title{Ordinal outcome tests}
\usage{
ordinal_tests(x, y, sample_size, K)
}
\arguments{
\item{y}{}
}
\description{
(Description. Mention the individual tests used here.)
}
