% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_ratio_text}
\alias{parse_ratio_text}
\title{Parse Ratio Text}
\usage{
parse_ratio_text(text)
}
\arguments{
\item{text}{A string of in the form of e.g. 5:95 or 70:30}
}
\value{
Numeric vector of length 2
}
\description{
This function parses text from ratios which are written in the format of 1-2 digit numbers separated by a colon and trailing with another 1-2 digit number. The text is processed into a numeric vector of length 2 containing the two numbers.
}
\examples{

parse_ratio_text("70:30")

}
