% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_distribution_results}
\alias{plot_distribution_results}
\title{Plot Distribution}
\usage{
plot_distribution_results(df, alpha = 0.05, outlier_removal = 0.1)
}
\arguments{
\item{df}{data frame of p-values}

\item{alpha}{significance level}

\item{outlier_removal}{top x proportion of observations to filter out of table}
}
\value{
ggplot object
}
\description{
This function takes a wide table of p-values (i.e. one column for each statistical test), converts it to long format, and creates a density plot of the p-values by each test.
}
