% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_data_one_group_formatted}
\alias{simulation_data_one_group_formatted}
\title{Formatted Simulation Data for One Group}
\format{
## `simulation_data_one_group_formatted`
A list
\describe{
  \item{p_values}{A data frame of p-values from each run of each test}
  \item{initial_groups}{A nested list with information for each simulation run}
}
}
\usage{
simulation_data_one_group_formatted
}
\description{
Simulated p-values and metadata for a two group comparison in a tidy format. Useful for Type I error calculations.
}
\keyword{datasets}
