% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_data_two_groups}
\alias{simulation_data_two_groups}
\title{Simulation Data for Two Groups}
\format{
## `simulation_data_two_groups`
A list
\describe{
  \item{p_values}{A data frame of p-values from each run of each test}
  \item{initial_groups}{A nested list with information for each simulation run}
}
}
\usage{
simulation_data_two_groups
}
\description{
Simulated p-values and metadata for a two group comparison. Useful for Type II error and power calculations.
}
\keyword{datasets}
