% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticParameters-package.R
\docType{package}
\name{SyntheticParameters-package}
\alias{SyntheticParameters}
\alias{SyntheticParameters-package}
\title{SyntheticParameters: Ordinal Endpoint Comparisons}
\description{
Simultaneously evaluate multiple ordinal outcome measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NeuroShepherd/SyntheticParameters}
  \item Report bugs at \url{https://github.com/NeuroShepherd/SyntheticParameters/issues}
}

}
\author{
\strong{Maintainer}: Pat Callahan \email{patricktcallahan18@gmail.com} (\href{https://orcid.org/0000-0003-1769-7580}{ORCID})

}
\keyword{internal}
