% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_groups.R
\name{assign_groups}
\alias{assign_groups}
\title{Randomly assign groups}
\usage{
assign_groups(sample_size, sample_prob, prob0, prob1, seed)
}
\arguments{
\item{sample_size}{total number of people under observation.}

\item{prob0}{vector probability of each possible outcome for the null group}

\item{prob1}{vector probability of each possible outcome for the intervention group}

\item{seed}{integer specifying the seed number}

\item{K}{number of possible outcome values in the target variable}
}
\value{
list of group assignments
}
\description{
(Brief description of the function here.)
}
