% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tests.R
\name{ordinal_tests}
\alias{ordinal_tests}
\title{Ordinal outcome tests}
\usage{
ordinal_tests(x, y, ...)
}
\arguments{
\item{x}{Group one}

\item{y}{Group two}

\item{...}{Placeholder for additional arguments to functions}
}
\value{
A named vector of probabilities for each test
}
\description{
A metafunction that runs the statistical tests listed below, and returns the p-values as a named vector.
}
\details{
* `stats::wilcox.test()`
* `stats::fisher.test()`
* `stats::chisq.test(correct = FALSE)`
* `stats::chisq.test(correct = TRUE)`
* `rms::lrm()`
* `stats.kruskal.test()`
* `coin::independence_test(ytrafo = coin::rank_trafo)`
}
