% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalsimr-package.R
\docType{package}
\name{ordinalsimr-package}
\alias{ordinalsimr}
\alias{ordinalsimr-package}
\title{ordinalsimr: Compare Ordinal Endpoints Using Simulations}
\description{
Simultaneously evaluate multiple ordinal outcome measures. Applied data analysts in particular are faced with uncertainty in choosing appropriate statistical tests for ordinal data. The included 'shiny' application allows users to simulate outcomes given different ordinal data distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NeuroShepherd/ordinalsimr}
  \item \url{https://neuroshepherd.github.io/ordinalsimr/}
  \item Report bugs at \url{https://github.com/NeuroShepherd/ordinalsimr/issues}
}

}
\author{
\strong{Maintainer}: Pat Callahan \email{patricktcallahan18@gmail.com} (\href{https://orcid.org/0000-0003-1769-7580}{ORCID}) [copyright holder]

}
\keyword{internal}
