% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_power}
\alias{plot_power}
\title{Plot Test Power}
\usage{
plot_power(df, power_threshold = 0.8)
}
\arguments{
\item{df}{a dataframe with p-values and a sample_size column}

\item{power_threshold}{numeric. desired power threshold}
}
\value{
ggplot object
}
\description{
Plot Test Power
}
